# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BARKSDALE MUSIC GROUP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /collab/rooms:
    get:
      operationId: getCollabRooms
      tags: [collab]
      summary: List active collab rooms
      responses:
        "200":
          description: List of rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CollabRoom"
    post:
      operationId: createCollabRoom
      tags: [collab]
      summary: Create a live session room
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollabRoomInput"
      responses:
        "201":
          description: Created room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollabRoom"
  /collab/rooms/{id}:
    get:
      operationId: getCollabRoom
      tags: [collab]
      summary: Get session room detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Room detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollabRoom"
  /collab/skill-test:
    post:
      operationId: submitSkillTest
      tags: [collab]
      summary: Submit skill test results
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkillTestInput"
      responses:
        "200":
          description: Skill profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillProfile"
...
  /collab/battles:
    get:
      operationId: getBattles
      tags: [collab]
      summary: Get active beat battles
      responses:
        "200":
          description: List of battles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Battle"
    post:
      operationId: createBattle
      tags: [collab]
      summary: Start a beat battle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BattleInput"
      responses:
        "201":
          description: Created battle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Battle"
  /collab/battles/{id}/vote:
    post:
      operationId: voteBattle
      tags: [collab]
      summary: Vote on a battle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BattleVoteInput"
      responses:
        "200":
          description: Vote recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Battle"
  /collab/leaderboard:
    get:
      operationId: getLeaderboard
      tags: [collab]
      summary: Global producer leaderboard
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Leaderboard
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"
  /collab/micro-gigs:
    get:
      operationId: getMicroGigs
      tags: [collab]
      summary: List available micro-gig sessions
      responses:
        "200":
          description: Micro gigs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MicroGig"
    post:
      operationId: createMicroGig
      tags: [collab]
      summary: List your session time for sale
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MicroGigInput"
      responses:
        "201":
          description: Micro gig created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MicroGig"
...
components:
  schemas:
...
    CollabRoom:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [public, private, invite]
        roles:
          type: array
          items:
            type: object
            properties:
              role:
                type: string
              userId:
                type: ["string", "null"]
              username:
                type: ["string", "null"]
            required:
              - role
        participants:
          type: integer
        isLive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - type
        - roles
        - participants
        - isLive
    CollabRoomInput:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [public, private, invite]
        password:
          type: ["string", "null"]
        role:
          type: string
      required:
        - name
        - type
        - role
    SkillTestInput:
      type: object
      properties:
        scores:
          type: object
          properties:
            drums:
              type: integer
            bass:
              type: integer
            keys:
              type: integer
            sampling:
              type: integer
            arrangement:
              type: integer
            mixing:
              type: integer
            melody:
              type: integer
          required:
            - drums
            - bass
            - keys
            - sampling
            - arrangement
            - mixing
            - melody
      required:
        - scores
    SkillProfile:
      type: object
      properties:
        drums:
          type: integer
        bass:
          type: integer
        keys:
          type: integer
        sampling:
          type: integer
        arrangement:
          type: integer
        mixing:
          type: integer
        melody:
          type: integer
        strengths:
          type: array
          items:
            type: string
        weaknesses:
          type: array
          items:
            type: string
      required:
        - drums
        - bass
        - keys
        - sampling
        - arrangement
        - mixing
        - melody
        - strengths
        - weaknesses
...
    Battle:
      type: object
      properties:
        id:
          type: string
        prompt:
          type: string
        chords:
          type: string
        challenger:
          type: string
        opponent:
          type: ["string", "null"]
        status:
          type: string
          enum: [open, active, voting, complete]
        timeRemaining:
          type: ["integer", "null"]
        votesA:
          type: integer
        votesB:
          type: integer
        winner:
          type: ["string", "null"]
        expiresAt:
          type: string
      required:
        - id
        - prompt
        - chords
        - challenger
        - status
        - votesA
        - votesB
        - expiresAt
    BattleInput:
      type: object
      properties:
        prompt:
          type: string
        opponentId:
          type: ["string", "null"]
      required:
        - prompt
    BattleVoteInput:
      type: object
      properties:
        vote:
          type: string
          enum: [a, b]
      required:
        - vote
    LeaderboardEntry:
      type: object
      properties:
        rank:
          type: integer
        userId:
          type: string
        username:
          type: string
        score:
          type: integer
        category:
          type: string
        badges:
          type: array
          items:
            type: string
        rating:
          type: number
      required:
        - rank
        - userId
        - username
        - score
        - category
    MicroGig:
      type: object
      properties:
        id:
          type: string
        sellerId:
          type: string
        sellerName:
          type: string
        title:
          type: string
        description:
          type: string
        pricePerQuarterHour:
          type: number
        pricePerHour:
          type: number
        rating:
          type: number
        reviewCount:
          type: integer
        isAvailable:
          type: boolean
        skills:
          type: array
          items:
            type: string
      required:
        - id
        - sellerId
        - sellerName
        - title
        - pricePerQuarterHour
        - pricePerHour
        - isAvailable
    MicroGigInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        pricePerQuarterHour:
          type: number
        pricePerHour:
          type: number
        skills:
          type: array
          items:
            type: string
      required:
        - title
        - pricePerQuarterHour
        - pricePerHour
...