# lib/api-zod/src/generated/types/microGigInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * BARKSDALE MUSIC GROUP API
 * OpenAPI spec version: 0.1.0
 */

export interface MicroGigInput {
  title: string;
  description?: string;
  pricePerQuarterHour: number;
  pricePerHour: number;
  skills?: string[];
}